requirements.txt
azure-identity
azure.mgmt.confidentialledger
azure.confidentialledger